added a unit that will always be spawned at the start of the game for testing purposes in gamelogic. fixed a iteration error in compview

+++ include/ActorInterface.hpp
@@ -84,7 +84,7 @@ public:
 
     @return an integer representing the actor's identity
    */
-  int getID(){return id;}
+  long long getID(){return id;}
 
   /*
     Returns the actor's sprite.
